# openapi.yaml
openapi: 3.0.3
info:
  title: h26_07 API
  version: 0.1.0
  description: |
    通常数字球だけを扱う Web ビンゴアプリの HTTP API と WebSocket event body schema。
    ゲームルールの意味論は docs/domain.md、通信境界の設計は docs/api.md を正本とする。
servers:
  - url: http://localhost:8080
    description: 開発環境
security:
  - forwardedUser: []
...
paths:
...
  /api/rooms:
    get:
      tags:
        - rooms
      summary: ルーム一覧を取得
      description: finished を含む全ルームを返す。
      operationId: listRooms
      responses:
        "200":
          description: ルーム一覧。finished を含む。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - rooms
      summary: ルームを作成
      description: 認証済みユーザーがルームを作成する。作成者はそのルームの admin になる。
      operationId: createRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "200":
          description: 作成されたルーム。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/rooms/{roomId}:
    get:
      tags:
        - rooms
      summary: ルーム詳細を取得
      description: participant または admin は finished の部屋も取得できる。
      operationId: getRoom
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: ルーム詳細。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: このルームを閲覧する権限がない。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/RoomNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    forwardedUser:
      type: apiKey
      in: header
      name: X-Forwarded-User
      description: 認証済みユーザーの traP ID。
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      description: ルーム ID。
      schema:
        $ref: "#/components/schemas/RoomId"
  responses:
    BadRequest:
      description: リクエスト body が不正。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 認証されていない。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RoomNotFound:
      description: 指定したルームが存在しない。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: サーバ内部エラー。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    UUID:
      type: string
      description: UUID string。
    DateTime:
      type: string
      format: date-time
      description: ISO 8601 datetime string。
    UserId:
      type: string
      description: traP ID。
    RoomId:
      allOf:
        - $ref: "#/components/schemas/UUID"
      description: ルーム ID。
    RoomCode:
      type: string
      pattern: "^\\d{6}$"
      description: 6 桁の数字文字列。
      example: "123456"
    RoomState:
      type: string
      enum:
        - waiting
        - playing
        - finished
    PickState:
      type: string
      enum:
        - idle
        - picking
        - exhausted
...
    Error:
      type: object
      additionalProperties: false
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      additionalProperties: false
      properties:
        userId:
          $ref: "#/components/schemas/UserId"
      required:
        - userId
    GameSettingsInput:
      type: object
      additionalProperties: false
      description: 作成・更新時に渡すゲーム設定。adminUserIds は指定した場合だけ admin 一覧を置き換える。
      properties:
        name:
          type: string
        description:
          type: string
        adminUserIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/UserId"
      required:
        - name
        - description
    GameSettings:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        description:
          type: string
        admins:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/User"
      required:
        - name
        - description
        - admins
    ParticipantSummary:
      type: object
      additionalProperties: false
      description: 参加者一覧や display 向けの概要。admin 権限、カード詳細、リーチ情報、ビンゴ情報は含めない。
      properties:
        user:
          $ref: "#/components/schemas/User"
        joinedAt:
          $ref: "#/components/schemas/DateTime"
      required:
        - user
        - joinedAt
    Room:
      type: object
      additionalProperties: false
      properties:
        roomId:
          $ref: "#/components/schemas/RoomId"
        roomCode:
          $ref: "#/components/schemas/RoomCode"
        state:
          $ref: "#/components/schemas/RoomState"
        pickState:
          $ref: "#/components/schemas/PickState"
        qrCodeVisible:
          type: boolean
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ParticipantSummary"
        bingoSummaries:
          type: array
          items:
            $ref: "#/components/schemas/BingoSummary"
        reachSummaries:
          type: array
          items:
            $ref: "#/components/schemas/ReachSummary"
        settings:
          $ref: "#/components/schemas/GameSettings"
        createdAt:
          $ref: "#/components/schemas/DateTime"
        updatedAt:
          $ref: "#/components/schemas/DateTime"
      required:
        - roomId
        - roomCode
        - state
        - pickState
        - qrCodeVisible
        - participants
        - bingoSummaries
        - reachSummaries
        - settings
        - createdAt
        - updatedAt
...
    BingoSummary:
      type: object
      additionalProperties: false
      description: 現在ビンゴしている参加者の概要。
      properties:
        user:
          $ref: "#/components/schemas/User"
        bingoOrders:
          type: array
          items:
            type: integer
            minimum: 1
      required:
        - user
        - bingoOrders
    ReachSummary:
      type: object
      additionalProperties: false
      description: 現在リーチしていて、まだビンゴしていない参加者の概要。
      properties:
        user:
          $ref: "#/components/schemas/User"
      required:
        - user
...
    CreateRoomRequest:
      type: object
      additionalProperties: false
      properties:
        settings:
          $ref: "#/components/schemas/GameSettingsInput"
      required:
        - settings
...